Replace `dtype` parameter in archives with `solution_dtype`, `objective_dtype`, and `measures_dtype` (#639)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Resolves #638

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Replace current `parse_dtype` util -- note that it takes in an
`xp` parameter that is currently set to `np_compat` across the archives;
eventually, it should be the `xp` taken in by the archive.
- [x] Update each archive by doing the following:
  - Mark `dtype` as deprecated
  - Add `solution_dtype`, `objective_dtype`, and `measures_dtype`
  - Add error if `dtype` is passed in
  - Update dtype calls in ArrayStore
- [x] Remove `ArchiveDType` from `ribs.typing`
- [x] Update `dtypes` annotations for archives and ArrayStore
- [x] Migrate current usages of `dtype`

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have linted and formatted my code with `ruff` and `ty`
- [x] I have tested my code by running `pytest`
- [x] I have added a description of my change to the changelog in
`HISTORY.md`
- [x] This PR is ready to go

+++ tests/archives/density_archive_test.py
@@ -49,7 +49,7 @@ def test_density_dtype(density_method, dtype):
         buffer_size=10000,
         density_method=density_method,
         bandwidth=2.0,
-        dtype=dtype,
+        measures_dtype=dtype,
     )
 
     measures = np.arange(10).reshape((5, 2))
